Implementation in Object_Calculate_FWHM() of a brute force method of fitting a Moffat curve to the radial profile of each object. Manual testing on a few test objects shows it fits well to the data and produces a FWHM within 0.2 pixels (0.05") of that calculated by IRAF.
__Radial profile__ entails converting x,y,z -> r,z where r = sqrt((x-xpos)^2 + (y-ypos)^2), i.e. based
around the barycentre (1st moment). Using a gsl routine for this as it executes a little faster than
standard sqrt math lib function and we have a lot of pixels to work on.

__Moffat curve__ is a better fit to a stellar radial profile than a gaussian, and is given by the equation
y = k * [ 1 + (x/a)^2 ]^(-b), where k = peak at x=0 and a,b are parameters to find. 'k' can be set to the
nearest (brightest) pixel to the barycentre as it's close enough.

This version (coded in December 2007) uses a crude scan through a range of a,b values to find the best fit,
purely to enable testing on a large number of objects in many frames.